pyproject.toml
[tool.poetry]
name = "my-rag-project"
version = "0.1.0"
description = "Enterprise RAG with LlamaIndex and LangGraph"
authors = ["Your Name <your@email.com>"]
packages = [{include = "rag", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

# --- LlamaIndex 核心 ---
# 必须使用 0.13.0+ 以适配最新的 Qdrant 插件功能
llama-index-core = "^0.13.0"

# --- 阿里云 DashScope 组件 ---
llama-index-llms-dashscope = ">=0.2.0"
llama-index-embeddings-dashscope = ">=0.2.0"
llama-index-postprocessor-dashscope-rerank = ">=0.2.0"

# --- OpenAI 兼容层 ---
llama-index-embeddings-openai = ">=0.2.0"

# --- 向量数据库 ---
# 关键：修复 QdrantClient 调用问题的版本
llama-index-vector-stores-qdrant = "^0.9.1"
qdrant-client = "^1.16.2"

# --- 文件读取 ---
llama-index-readers-file = ">=0.2.0"
pypdf = "^6.0.0"

# --- 稀疏向量生成 (jieba 中文分词 + 哈希稀疏向量) ---
jieba = "^0.42.0"

# --- 核心框架 ---
langgraph = "^0.2.0"
pydantic-settings = "^2.0.0"

# --- LangChain 全家桶 (用于 Tool 封装) ---
langchain = "^0.2.0"
langchain-community = "^0.2.0"
langchain-core = "^0.2.0"
langchain-openai = "^0.1.0"

# --- 基础设施与 UI ---
gradio = "^5.0.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
transformers = ">=4.40.0"
typer = "^0.21.1"
fastapi = "^0.128.6"
uvicorn = "^0.40.0"
httpx = "^0.28.1"
pyyaml = "^6.0.3"
pandas = "^2.0.0"
numpy = "^1.24.0"
tqdm = "^4.66.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"